chore: lint

+++ graft/evm/firewood/triedb.go
@@ -123,7 +123,7 @@ func New(config Config) (*TrieDB, error) {
 	}
 
 	options := []ffi.Option{
-		ffi.WithNodeCacheEntries(uint(config.CacheSizeBytes / 256)), // TODO(#4750): is 256 bytes per node a good estimate?
+		ffi.WithNodeCacheEntries(config.CacheSizeBytes / 256), // TODO(#4750): is 256 bytes per node a good estimate?
 		ffi.WithFreeListCacheEntries(config.FreeListCacheEntries),
 		ffi.WithRevisions(config.RevisionsInMemory),
 		ffi.WithReadCacheStrategy(config.CacheStrategy),
